requirements.txt
absl-py==2.1.0
dynamic-pip==1.1.1
mpi4py==3.1.5
numpy==1.24.3
nvitop==1.3.1
openai==1.3.7
requests==2.31.0
